fixed: 좋아요 추가 및 삭제 방법 수정

+++ src/main/java/colony/webproj/controller/LikesController.java
@@ -19,24 +19,25 @@ public class LikesController {
     private final LikesService likesService;
 
     /**
-     *  좋아요 추가
+     *  좋아요 추가 및 삭제
      */
     @PostMapping("/heart/{answerId}")
     public ResponseEntity<?> increaseLikes(@PathVariable("answerId") Long answerId,
                                            @AuthenticationPrincipal PrincipalDetails principalDetails) throws Exception{
 
-        return likesService.addLikes(answerId, principalDetails.getLoginId()) ?
-                ResponseEntity.ok(true) : ResponseEntity.badRequest().body("false");
+        boolean check = likesService.getLike(answerId,principalDetails.getLoginId());
+
+        if(check){
+            likesService.addLikes(answerId,principalDetails.getLoginId());
+            return ResponseEntity.ok("add like");
+        }
+        else{
+            likesService.removeLikes(answerId, principalDetails.getLoginId());
+            return ResponseEntity.ok("remove like");
+
+        }
+
 
     }
-
-    /**
-     * 좋아요 취소
-     */
-    @DeleteMapping("/heart/{answerId}")
-    public ResponseEntity<?> decreaseLikes(@PathVariable("answerId") Long answerId,
-                                           @AuthenticationPrincipal PrincipalDetails principalDetails) throws Exception{
-        likesService.removeLikes(answerId, principalDetails.getLoginId());
-        return ResponseEntity.ok(true);
-    }
+    
 }
